tired from triggers api not working and create my own api :)

+++ app/src/main/java/ru/yellowshark/dsr_weather/data/remote/request/Area.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Area(
     @SerializedName("coordinates")
-    val coordinates: List<Int>,
+    val coordinates: List<Double>,
     @SerializedName("type")
     val type: String
 )
